+++ .changeset/light-roses-count.md
@@ -1,0 +1,5 @@
+---
+"@spectrum-css/preview": minor
+---
+
+Pull out decorators into their own files
